+++ lib/src/utils/code_builder.dart
@@ -1,0 +1,16 @@
+import 'package:code_builder/code_builder.dart';
+
+Block ifStatement(
+  Expression conditional,
+  Code body, {
+  bool addCurlyBraces = true,
+}) {
+  return Block.of([
+    Code('if ('),
+    conditional.code,
+    Code(')'),
+    if (addCurlyBraces) Code('{'),
+    body,
+    if (addCurlyBraces) Code('}'),
+  ]);
+}